Find if Flash and Java are the latest versions
Will now invoke a web request to the respective website to determine the
current version of Flash and Java. Implemented a basic form of error
handling. This will improve as I continue to work on the script.  For
PowerShell 2.0 environments, the latest version can be pulled from a
text file but this needs to be run on a computer with PowerShell 3.0 or
higher to generate.

Changed to 'Get-ItemProperty' from the shorthand 'gp' for readability.
Capitalized 'Where-Object' and 'ForEach-Object' for consistency.